split servers
