+++ apps/front/package.json
@@ -4,7 +4,7 @@
   "main": "src/index.tsx",
   "type": "module",
   "scripts": {
-    "dev": "node server.js",
+    "dev": "node server/server.dev.js",
     "build:client": "vite build --ssrManifest --outDir dist/client",
     "build:server": "vite build --ssr src/server/index-server.tsx --outDir dist/server",
     "build:scripts": "vite build -c vite.scripts.config.ts",
@@ -12,7 +12,7 @@
     "build": "npm run build:server && npm run build:client && npm run build:scripts && npm run build:static",
     "generate": "node dist/_scripts/exe-prerender.js",
     "preview": "serve dist/static",
-    "start": "cross-env NODE_ENV=production node server",
+    "start": "cross-env NODE_ENV=production node server/server.prod",
     "test:watch": "vitest",
     "test": "vitest run"
   },
